feat: Implement database connection and initialization
- Added MySQL and MongoDB connection setup in config/database.js.
- Created dbInit.js for initial data population in the database.
- Introduced multer.js for file upload handling.
- Developed authController.js for user authentication routes.
- Created bookingController.js for managing bookings.
- Implemented homeController.js for rendering the home page and property filtering.
- Added listingController.js for handling property listings.
- Created uploadController.js for managing file uploads.
- Developed userController.js for user profile and dashboard management.
- Refactored index.js to utilize route modules for better organization.
- Added middleware for authentication and session management.
- Created routes for authentication, booking, home, user, listing, and uploads.
