+++ routes/bookingRoutes.js
@@ -1,0 +1,8 @@
+const express = require("express");
+const router = express.Router();
+const bookingController = require("../controllers/bookingController");
+
+// Booking routes
+router.get("/:id", bookingController.bookListing);
+
+module.exports = router;
