+++ routes/userRoutes.js
@@ -1,0 +1,13 @@
+const express = require("express");
+const router = express.Router();
+const userController = require("../controllers/userController");
+const { getUserFromSession } = require("../middleware/auth");
+
+// Apply getUserFromSession middleware to all routes
+router.use(getUserFromSession);
+
+// User profile routes
+router.get("/profile/:userId", userController.getProfile);
+router.get("/dashboard/:userId", userController.getDashboard);
+
+module.exports = router;
